requirements.txt
openai==0.28.0
python-dotenv==1.0.0
streamlit==1.28.1
